Fix precision and signedness of integer literal types in Java. Resolves #120.

+++ hydra-java/src/main/java/hydra/lib/literals/Int8ToBigint.java
@@ -33,7 +33,7 @@ public class Int8ToBigint extends PrimitiveFunction {
         return args -> Flows.map(Expect.int8(args.get(0)), s -> Terms.bigint(apply(s)));
     }
 
-    public static BigInteger apply(Short value) {
+    public static BigInteger apply(Byte value) {
         return BigInteger.valueOf(value);
     }
 }
